third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEMorphologyElement-dom-in-attr-expected.txt

SVG 1.1 dynamic update tests
Erode radius = 4

Tests dynamic updates of the 'in' attribute of the SVGFEMorphologyElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS morphologyElement.getAttribute('in') is "SourceAlpha"
PASS morphologyElement.getAttribute('in') is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
